+++ activesupport/test/core_ext/module/attribute_accessor_per_thread_test.rb
@@ -6,17 +6,13 @@ require "active_support/core_ext/module/attribute_accessors_per_thread"
 class ModuleAttributeAccessorPerThreadTest < ActiveSupport::TestCase
   setup do
     @class = Class.new do
-      def self.name; "MyClass#{object_id}"; end
-
       thread_mattr_accessor :foo
       thread_mattr_accessor :bar,  instance_writer: false
       thread_mattr_reader   :shaq, instance_reader: false
       thread_mattr_accessor :camp, instance_accessor: false
     end
 
-    @subclass = Class.new(@class) do
-      def self.name; "Sub#{super}"; end
-    end
+    @subclass = Class.new(@class)
 
     @object = @class.new
   end
@@ -62,15 +58,20 @@ class ModuleAttributeAccessorPerThreadTest < ActiveSupport::TestCase
     end.join
   end
 
-  def test_default_value_is_the_same_object
-    default = Object.new
+  def test_nonfrozen_default_value_is_duped_and_frozen
+    default = []
+    @class.thread_mattr_accessor :baz, default: default
+
+    assert_equal default, @class.baz
+    assert_predicate @class.baz, :frozen?
+    assert_not_predicate default, :frozen?
+  end
+
+  def test_frozen_default_value_is_not_duped
+    default = [].freeze
     @class.thread_mattr_accessor :baz, default: default
 
     assert_same default, @class.baz
-
-    Thread.new do
-      assert_same default, @class.baz
-    end.join
   end
 
   def test_should_use_mattr_default