+++ railties/lib/rails/application/configuration.rb
@@ -298,6 +298,7 @@ module Rails
           if respond_to?(:active_support)
             active_support.default_message_encryptor_serializer = :json
             active_support.default_message_verifier_serializer = :json
+            active_support.raise_on_invalid_cache_expiration_time = true
           end
 
           if respond_to?(:action_controller)